Delete __pycache__ directory
Deleted Pychache directory. This directory is created each time we run the code in vscode to optimize future runs. It does not effect the functionality of our code in any way and should not be tracked. I created a .gitignore file to prevent github from tracking files like this in the future.